fixing build conflicts
moving `toActionName` to `flow.cpp`
